REFACT: Split out `stringBuilder.h/.c` and `storedBinaryValue.h/.c`.
